Add files via upload
This Zip File Contains:

1-Covid-19 Report MS Word (i191207) : 
Contains the report in MS Word format

2-Covid-19 Report PDF (i19-1207) : 
Contains the same report in PDF format

3-COVID-19 (i19-1207 Code File).ipynb :
Requires Anaconda to run.
It is a Jupyter Notebook file containing the code. All the cells can be run in one go to generate the outputs.
This file includes all the required tasks in detail.
The sections are clearly defined for each task along with the comments.
Main Libraries Required:
pystan, fbprophet, seaborn, pandas, numpy, matplotlib, plotly


4-COVID-19.csv : 
Data
The code inside the Jupyter Notebook is using this name to access the data. 
Please make sure to place this file in the same directory as "COVID-19 (i19-1207 Code File).ipynb"